Refactoring how we "demand" the outfile_extension
Prior to this commit, only the base_generator was allowed to skip the
outfile_extension.  Something of a code smell, but here we are.

With this commit, to address the emergence of the CopyGenerator, I need
to allow for both the nil default output_extension as well as how we'd
go about writing that file to the correct output location.

(I'm tired, naming things is hard)