Using template for CopyGenerator
As part of prior pairing sessions, we wanted to incorporate the idea of
a template for the path.

With this change, I'm testing the behavior within the confines of the
CopyGenerator.  Mostly as a proof of concept.

The next part is to roll this out to the PdfSplitGenerator; then
generalize and finally tidy-up methods no longer used.

Related to:

- https://github.com/scientist-softserv/derivative_rodeo/issues/2