♻️ Renaming "target" to "location"
> What's in a name? That which we call a rose /
> By any other name would smell as sweet.
>
> --- William Shakespeare, *Romeo and Juliet*

We've been iterating through the naming of these concepts; the word
target made more sense than it's predecessor; a name now lost amongst
the sands of time.

With the change to "location" we're hoping that this helps further
clarify the sense of named place.  We can think of this as storage, but
storage comes with an idea of permanance and anything in an SQSLocation
is by it's nature ephmeral.

Closes: #15

- https://github.com/scientist-softserv/derivative_rodeo/issues/15