♻️ Refactoring BaseLocation to initialize with config
In my reading of the SqsLocation class, I saw repeated configuration
information.  This felt like something I could move elsewhere; and
expose in a more general sense.

A future Jeremy is considering how different locations would have
configurations and that those may not be something that is appropriate
to include in the base Configuration class, but instead allow for
plugins.

But that is for the future.