🎁 WIP - Adding Thumbnail Generator skeleton
This commit introduces the skeleton of the thumbnail generation
necessary for PDFs, images, and the images that we make from the PDFs.

It does not have the full business logic to perform this work; but the
specs are mostly in place and can then be used to drive development.

Related to #22

- https://github.com/scientist-softserv/derivative_rodeo/issues/22