♻️ Improving Legibility of PDF Splitting
As part of working on #219, I was looking into "What is the pathing
convention for the split images?"  As I was reading the code to get to
that answer, I realized that I wanted to see that information in the
`DerivativeRodeo::Generators::PdfSplitGenerator`.

Further, given that the PdfSplitGenerator is part of the more visible
"interface" of the DerivativeRodeo, I think it is important for the
filename to be present.

Thus I moved towards this refactor, to better expose that and remove
additional chatter in the code.  There are still a few things remaining
but this has helped re-inforce my understanding of the PDF Splitting.

I could further "compress" the basename template concept into the fully
qualitified filename; but I think that would confuse concepts.

Related to:

- https://github.com/scientist-softserv/iiif_print/issues/219