🎁 Adding PDF Split Page Checks
Prior to this commit, if we'd already pre-processed a PDF split, we
would again re-process that split (as there was no check for existing
pages).

With this commit, we check for those pre-processed pages.

One critical bit of conversation, is that one work might have multiple
PDFs uploaded.  Therefore, it is important to have those PDFs pages
write to different "sub-directories".  I'm putting this hear so we can
account for that in a test audit of some kind.

Related to:

- https://github.com/scientist-softserv/adventist-dl/issues/330
- https://github.com/scientist-softserv/iiif_print/issues/220

Co-authored-by: Rob Kaufman <rob@notch8.com>
Co-authored-by: Kirk Wang <k3wang@gmail.com>