🎁 Adding PDF Split Page Checks (#36)
* 🎁 Adding PDF Split Page Checks

Prior to this commit, if we'd already pre-processed a PDF split, we
would again re-process that split (as there was no check for existing
pages).

With this commit, we check for those pre-processed pages.

One critical bit of conversation, is that one work might have multiple
PDFs uploaded.  Therefore, it is important to have those PDFs pages
write to different "sub-directories".  I'm putting this hear so we can
account for that in a test audit of some kind.

Related to:

- https://github.com/scientist-softserv/adventist-dl/issues/330
- https://github.com/scientist-softserv/iiif_print/issues/220

Co-authored-by: Rob Kaufman <rob@notch8.com>
Co-authored-by: Kirk Wang <k3wang@gmail.com>

* ☑️ Verifying pdf splitter finds pre-existing files

Updating a bit of documentation and reworking the filename to account
for a work having multiple PDFs.

- https://github.com/scientist-softserv/adventist-dl/issues/330
- https://github.com/scientist-softserv/iiif_print/issues/220

* ☑️ Refining globbed_tail_locations for S3

Prior to this commit, we didn't have a spec for the S3 behavior.  We now
have a test for an S3 Faux Bucket.

Related to:

- https://github.com/scientist-softserv/adventist-dl/issues/330
- https://github.com/scientist-softserv/iiif_print/issues/220

---------

Co-authored-by: Rob Kaufman <rob@notch8.com>
Co-authored-by: Kirk Wang <k3wang@gmail.com>