♻️ Leveraging Marcel to better derive "mime_type"
Prior to this we were using a naive extension sniffer to determine a
"mime_type" (which we used to decide on thumbnail dimensions).

With this commit, we leverage Marcel (a Rails gem) that can detect mime
based on the magic stings at the beginning of files or based on
extensions and some lower "cost" processing (compared to `fits.sh`).

Related to:

- https://github.com/scientist-softserv/derivative_rodeo/issues/30
- https://github.com/scientist-softserv/adventist-dl/issues/330