♻️ Flattening assumptive dir structure for PDFs
Yes, it would be nice to have sub-directories for pages ripped from a
PDF.  However, that aspirational state creates complications on
the implementation details of the DerivativeRodeo; by moving from a tail
glob to a regular expression, we create a more powerful mechanism for
finding files.

These changes also highlighted a few implementation bugs (namely
ensuring the correct expected return value of the newly re-named
function).

In IIIF Print we still need to consider how to find the child page's
derivatives of an original PDF.  That is, however, not a problem for
this repository.

Related to:

- https://github.com/scientist-softserv/iiif_print/issues/251

Co-authored-by: Kirk Wang <kirk.wang@scientist.com>