🐛 Exposing means for generator to specify output extension
Prior to this commit, we did not account for a generator having a stake
in the output URI; that is the extension specified in the generator.

The behavior we saw was that our lambdas completed successfully for the
split PDF, OCR, etc.  However, the returned URIs were not the derivative
file (e.g. `filename--page-1.thumbnail.jpeg`) but were instead the input
file (e.g. `filename--page-1.tiff`) repeated back as output.

With this change, the output and pre-processed template can use the
suggested output extension of the generator.

Note: there are probably bugs for cases of SAME extension.  We'll work
around that.