🐛 Hacking away the .mono suffix for 2nd order derivatives (#53)
Given the "20121816.ARCHIVAL--page-3.tiff" page, prior to this commit we
were generating the following files:

- 20121816.ARCHIVAL--page-3.mono.alto.xml
- 20121816.ARCHIVAL--page-3.mono.coordinates.json
- 20121816.ARCHIVAL--page-3.mono.hocr
- 20121816.ARCHIVAL--page-3.mono.plain_text.txt
- 20121816.ARCHIVAL--page-3.mono.tiff
- 20121816.ARCHIVAL--page-3.thumbnail.jpeg

Iiif Print finds the derived files for "20121816.ARCHIVAL--page-3.tiff"
by looking for the basename (e.g. "20121816.ARCHIVAL--page-3") and then
the generator's output extension (e.g. 'plain_text.txt',
'coordinates.json', etc.)

With this change we chop out the ".mono" out of the locations and should
see the following:

- 20121816.ARCHIVAL--page-3.alto.xml
- 20121816.ARCHIVAL--page-3.coordinates.json
- 20121816.ARCHIVAL--page-3.hocr
- 20121816.ARCHIVAL--page-3.plain_text.txt
- 20121816.ARCHIVAL--page-3.mono.tiff
- 20121816.ARCHIVAL--page-3.thumbnail.jpeg