🎁 Updating logging to include class name
In reviewing the logs in SpaceStone, I can tell which handler was
running based on the lambda's name.  However that does not directly map
to the Derivative Rodeo generator.

In otherwords, prior to this commit the logging was not providing the
context of the generator.

With this commit, we're including that information as a means of helping
debug and triage.