🐛 Fixing incorrect method name
This commit includes 3 changes:

1. Replacing a misspelled name with the correct name
2. Improving logging
3. Adding a parameter to an exception.

At one point we had a method `globbed_tail_locations` however with some
refactoring I renamed that to `matching_locations_in_file_dir`; however
I missed the `globbed_tail_loations`.

In addition to fixing the misnamed method name, I'm adding some improved
logging that helps with a problem I've encountered.  Namely that I got a

`#<ArgumentError: invalid byte sequence in UTF-8>` on a string.  That
string was the contents of the file found at the `path_to_hocr`; however
the exception didn't provide insight into the filename.

With this change, I fix a method that was broken and improve logging.

Last, the exception for buckets used an unprovided variable (e.g.
`file_uri`).  This now provides that expected file_uri.

Related to:

- https://github.com/scientist-softserv/derivative_rodeo/issues/56