🎁 Adding logging for PDF splitting information
Prior to this commit, during run time we didn't have much insight as to
whether or not we were finding the split pages (and if so how many).

With this commit we log information regarding the number of files we
find, the path/index for each one OR if we're generating our own.

Related to:

- https://github.com/scientist-softserv/derivative_rodeo/issues/56