🐛 Correcting preprocessed_location_template for PDF
Prior to this commit the `preprocessed_location_template` was the original
PDF's location (as in "s3://bucket/aark_id/filename.ARCHIVAL.pdf").
This is because the calling of the generator has one file (e.g. the PDF)
and multipe output files (e.g. the pages of the PDF).

What we ideally want is for the `preprocessed_location_template` to
handle pages.  To know what page we're on requires knowledge of the
`input_location`, and the convention that each "page" ends with the
suffix `--page-NUMBER.EXTENSION`.

With this commit, we provide a crease for coercing the template into a
page specific format.  If we don't have this change, then we end up
using the `filename.ARCHIVAL.pdf` as the "tiff" for each of the pages.
This results in very odd behavior, very much unexpected.

Related to:

- https://github.com/scientist-softserv/derivative_rodeo/issues/56