🐛 Copying pre-processed files to output location
The following user-story reflects the desired behavior:

```gherkin
Given an input uri
When I call #generated_files
Then there will exist a corresponding file at the template-derived output location
```

Prior to this commit, if a file did not exist at the output location but
did exist at the pre-processed location, we would return the
pre-processed location (which existed).

However, the goal of the `#generated_files` method was to ensure that
there were files at the output location based on the input URIs and the
output and pre-processed templates.

With this commit, we have a scenario

```gherkin
Given an input uri
And a corresponding template-derived output location that does not exist
And a corresponding template-derived preprocessed location that does exist
When I call #generated_files
Then the we will copy the preprocessed location to the output location
And thus fulfill the contract of #generated_files
```

Related to:

- https://github.com/scientist-softserv/derivative_rodeo/issues/56