🐛 Add missing `#matching_locations_in_file_dir` method
In working towards gracefully handling a failed copy of a resource in
SpaceStone, we conditionally updated the `preprocess_template_location` to
be the original HTTP(s) location.  The idea being that if it does not
exist in the pre-processed location

The
[DerivativeRodeo::Generators::PdfSplitGenerator](https://github.com/scientist-softserv/derivative_rodeo/blob/84aa08e8387f105488a2b1a1e3821abacff4ad64/lib/derivative_rodeo/generators/pdf_split_generator.rb#L70-L80)
method (see below) can gracefully handle returning an empty list of
output locations.

```ruby
def existing_page_locations(input_location:)
  # See image_file_basename_template
  tail_regexp = %r{#{input_location.file_basename}--page-\d+\.#{output_extension}$}

  output_locations = input_location.derived_file_from(template: output_location_template).matching_locations_in_file_dir(tail_regexp: tail_regexp)
  return output_locations if output_locations.count.positive?

  return [] if preprocessed_location_template.blank?

  input_location.derived_file_from(template: preprocessed_location_template).matching_locations_in_file_dir(tail_regexp: tail_regexp)
end
```

At issue is understanding if we need to use the copy task to copy the
file locally.  We shall see once this is in play.

Related to:

- https://github.com/scientist-softserv/iiif_print/pull/283
- https://github.com/scientist-softserv/derivative_rodeo/issues/71