🐛 Improve handling of pre-requisite templates
Prior to this commit, we made a non-general assumption about the
template structure of the `HocrGenerator` `output_location_template`
structure that works for SpaceStone, but was not working for IiifPrint.

Namely there was a naive assumption that the template would have
extension information; and we could simply gsub the file extension for
another.  This worked with SpaceStone, because it did not provide a file
extension.  It failed with IiifPrint because the template had a file
extension, but that extension was different than the configuration.

The reason for the differences is that SpaceStone wants to write to a
consistent location; and the template process works for that.  But the
IiifPrint is need to write to a pre-defined location (per Hyrax's
implementation details).

By moving away from the hard-coded assumption of extensions, and
leveraging the location templates we maintain a greater degree of
flexibility.

Related to:

- https://github.com/scientist-softserv/adventist-dl/issues/695